Store DoorBoards objects to the server

+++ server/src/main/java/umm3601/note/NoteController.java
@@ -27,6 +27,7 @@ import io.javalin.http.NotFoundResponse;
 import io.javalin.http.UnauthorizedResponse;
 import umm3601.JwtProcessor;
 import umm3601.UnprocessableResponse;
+import umm3601.doorboard.DoorBoard;
 
 
 /**
@@ -43,6 +44,7 @@ public class NoteController {
   JacksonCodecRegistry jacksonCodecRegistry = JacksonCodecRegistry.withDefaultObjectMapper();
 
   private final MongoCollection<Note> noteCollection;
+  private final MongoCollection<DoorBoard> doorBoardCollection;
 
   /**
    * @param database the database containing the note data
@@ -53,10 +55,31 @@ public class NoteController {
       DeathTimer dt,
       JwtProcessor jwtProcessor) {
     jacksonCodecRegistry.addCodecForClass(Note.class);
+    jacksonCodecRegistry.addCodecForClass(DoorBoard.class);
+
     noteCollection = database.getCollection("notes").withDocumentClass(Note.class)
         .withCodecRegistry(jacksonCodecRegistry);
+
+    doorBoardCollection = database.getCollection("doorBoards").withDocumentClass(DoorBoard.class)
+        .withCodecRegistry(jacksonCodecRegistry);
+
     deathTimer = dt;
     this.jwtProcessor = jwtProcessor;
+  }
+
+  /**
+   * Given the ObjectId, as a hex string, of a DoorBoard in the database,
+   * return that DoorBoard.
+   *
+   * Returns null if the ObjectId doesn't correspond to any DoorBoard in the
+   * database.
+   *
+   * Throws an IllegalArgumentException if doorBoardID isn't a valid ObjectId.
+   */
+  private DoorBoard getDoorBoard(String doorBoardID) {
+    return doorBoardCollection
+      .find(eq("_id", new ObjectId(doorBoardID)))
+      .first();
   }
 
   /**
@@ -71,22 +94,28 @@ public class NoteController {
   public void deleteNote(Context ctx) {
     String id = ctx.pathParam("id");
 
-    String doorBoardID = ctx.queryParam("doorBoardid");
-    Note note;
+    // This throws an UnauthorizedResponse if the user isn't logged in.
+    String currentUserSub = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
 
+    Note note;
     try {
       note = noteCollection.find(eq("_id", new ObjectId(id))).first();
     } catch(IllegalArgumentException e) {
       throw new BadRequestResponse("The requested note id wasn't a legal Mongo Object ID.");
     }
+
     if (note == null) {
       throw new NotFoundResponse("The requested does not exist.");
-    } else if (note.doorBoardID != doorBoardID) {
-      throw new ForbiddenResponse("The requested note does not belong to this doorBoard. It cannot be deleted.");
-    } else {
-      noteCollection.deleteOne(eq("_id", new ObjectId(id)));
-      deathTimer.clearKey(id);
     }
+
+    String subOfOwnerOfNote = getDoorBoard(note.doorBoardID).sub;
+    if (!currentUserSub.equals(subOfOwnerOfNote)) {
+      throw new ForbiddenResponse("The requested note does not belong to this doorBoard. It cannot be deleted.");
+    }
+
+    noteCollection.deleteOne(eq("_id", new ObjectId(id)));
+    deathTimer.clearKey(id);
+    ctx.status(204);
   }
 
   /**
@@ -141,17 +170,24 @@ public class NoteController {
     }
 
     // For any other request, you have to be logged in.
-    String currentUserId;
+    String currentUserSub;
     try {
-      currentUserId = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
+      currentUserSub = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
     } catch (UnauthorizedResponse e) {
       throw e;
       // Catch and rethrow, just to be explicit.
     }
 
-    // You're only allowed to view your own notes.
-    if (!ctx.queryParamMap().containsKey("doorBoardid")
-        || !ctx.queryParam("doorBoardid").equals(currentUserId)) {
+    // You can't view everyone's notes (unless you're only asking for active
+    // notes.)
+
+    if (!ctx.queryParamMap().containsKey("doorBoardid")) {
+      throw new ForbiddenResponse(
+        "Request not allowed; users can only view their own notes.");
+    }
+
+    String subOfOwnerOfDoorBoard = getDoorBoard(ctx.queryParam("doorBoardid")).sub;
+    if (!currentUserSub.equals(subOfOwnerOfDoorBoard)) {
       throw new ForbiddenResponse(
         "Request not allowed; users can only view their own notes.");
     }
@@ -165,7 +201,7 @@ public class NoteController {
    */
   public void addNewNote(Context ctx) {
     Note newNote = ctx.bodyValidator(Note.class)
-      .check((note) -> note.doorBoardID == null) // The doorBoardID shouldn't be present; you can't choose who you're posting the note as.
+      .check((note) -> note.doorBoardID != null) // The doorBoardID shouldn't be present; you can't choose who you're posting the note as.
       .check((note) -> note.body != null && note.body.length() > 0) // Make sure the body is not empty
       .check((note) -> note.addDate != null && note.addDate.matches(ISO_8601_REGEX))
       .check((note) -> note.expireDate == null || note.expireDate.matches(ISO_8601_REGEX))
@@ -173,7 +209,25 @@ public class NoteController {
       .get();
 
     // This will throw an UnauthorizedResponse if the user isn't logged in.
-    newNote.doorBoardID = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
+    String currentUserSub = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
+
+
+    DoorBoard doorBoard;
+    try {
+      doorBoard = getDoorBoard(newNote.doorBoardID);
+    } catch (IllegalArgumentException e) {
+      throw new BadRequestResponse(
+        newNote.doorBoardID + "does not refer to an existing DoorBoard.");
+    }
+
+    if (doorBoard == null) {
+      throw new BadRequestResponse(
+        newNote.doorBoardID + "does not refer to an existing DoorBoard.");
+    }
+
+    if (!doorBoard.sub.equals(currentUserSub)) {
+      throw new ForbiddenResponse("You can only add notes to your own DoorBoard.");
+    }
 
     if(newNote.expireDate != null && !(newNote.status.equals("active"))) {
       throw new ConflictResponse("Expiration dates can only be assigned to active notices.");
@@ -220,9 +274,11 @@ public class NoteController {
     }
 
     // verifyJwtFromHeader will throw an UnauthorizedResponse if the user isn't logged in.
-    String currentUserID = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
+    String currentUserSub = jwtProcessor.verifyJwtFromHeader(ctx).getSubject();
 
-    if (!note.doorBoardID.equals(currentUserID)) {
+    String subOfOwnerOfNote = getDoorBoard(note.doorBoardID).sub;
+
+    if (!subOfOwnerOfNote.equals(currentUserSub)) {
       throw new ForbiddenResponse("Request not allowed; users can only edit their own notes");
     }
 
@@ -251,7 +307,7 @@ public class NoteController {
         if (validStatuses.contains(inputDoc.get("status"))) {
           toEdit.append("status", inputDoc.get("status"));
           noteStatus = inputDoc.get("status").toString();
-          if(inputDoc.get("status") != "active") {
+          if(!inputDoc.get("status").equals("active")) {
             toReturn.append("$unset", new Document("expireDate", ""));
             //Only active notices can have expiration dates, so if a notice becomes inactive, it loses
             //its expiration date.
@@ -265,7 +321,7 @@ public class NoteController {
       if(inputDoc.containsKey("expireDate")){
         if(inputDoc.get("expireDate") == null) {
           toReturn.append("$unset", new Document("expireDate", "")); //If expireDate is specifically included with a null value, remove the expiration date.
-        } else if (!(noteStatus.equals("active"))) {
+        } else if (!noteStatus.equals("active")) {
           throw new ConflictResponse("Expiration dates can only be assigned to active notices.");
           //Order of clauses means we don't mind of someone manually zeroes their expireDate when making something inactive.
         } else if(inputDoc.get("expireDate").toString() //This assumes that we're using the same string encoding they are, but it's our own API we should be fine.
